✨ feat(context): add cache persistence to local storage
