+++ tests/utils/cache.ts
@@ -1,4 +1,4 @@
-import { queryClient } from 'src/App'
+import queryClient from 'src/context/QueryClientProvider'
 
 export const invalidCache = () => {
   queryClient.clear()
